+++ mayan/apps/templating/classes.py
@@ -63,6 +63,7 @@ class Template:
                             'mathfilters.templatetags.mathfilters',
                             'mayan.apps.templating.templatetags.templating_datetime_tags',
                             'mayan.apps.templating.templatetags.templating_json_tags',
+                            'mayan.apps.templating.templatetags.templating_math_tags',
                             'mayan.apps.templating.templatetags.templating_regex_tags',
                             'mayan.apps.templating.templatetags.templating_tags'
                         ]